working / breaking on spartan

+++ R/get_travel_time.R
@@ -4,15 +4,23 @@ get_travel_time <- function(
   t_filename = "outputs/areaT.rds",
   tgc_filename = "outputs/areaTGC.rds",
   travel_time_filename = "outputs/travel_time.tif",
-  overwrite = TRUE
+  overwrite_raster = TRUE,
+  overwrite_t = TRUE
 ){
 
   npoints <- nrow(points)
 
   friction <- raster::raster(friction_surface)
 
-  tsn <- transition(friction, function(x) 1/mean(x), 8) # RAM intensive, can be very slow for large areas
-  saveRDS(tsn, t_filename)
+
+
+  if(!file.exists(t_filename) | (file.exists(t_filename) & overwrite_t)){
+    tsn <- transition(friction, function(x) 1/mean(x), 8) # RAM intensive, can be very slow for large areas
+    saveRDS(tsn, t_filename)
+  } else {
+    tsn <- readRDS(t_filename)
+  }
+
   tgc <- geoCorrection(tsn)
   saveRDS(tgc, tgc_filename)
